LPS-33867 Cleanup the interface, it is better to remove the getter/setter from functional contract interface
